+++ CHANGELOG.md
@@ -18,6 +18,7 @@ formatting guidelines.
 
 ### Added
 
+- MultitypeService now supports named dependencies.
 - `Resolver` now has a new requirement, `resolveAll(_:)`, which gives a mapping of name -> instance
   for all dependencies of a given type.
 - The factory now has a `clearDependencies()` method intended for testing.